backend/esp32_requirements.txt
pyserial==3.5
keyboard==0.13.5
requests==2.31.0
python-socketio[client]==5.8.0 